# openapi.yaml
openapi: 3.0.3
info:
  title: Aurora Backend API
  description: |
    Backend API untuk **Aurora Prediction Market** app.
    
    Dibangun dengan Express.js + TypeScript + PostgreSQL + Prisma ORM.
    Data market diambil dari **Polymarket API** (Gamma API, CLOB API, Data API).
    
    ## Autentikasi
    Sebagian besar endpoint user membutuhkan **JWT Bearer Token**.
    Login via `POST /api/auth/google` untuk mendapatkan token.
    
    ## Demo Mode
    Deposit dan Withdrawal bersifat **simulated** — tidak ada transaksi uang nyata.
    Setiap user baru mendapat $1000 USDC demo balance.
  version: 1.0.0
  contact:
    name: Aurora App
servers:
  - url: http://localhost:3000
    description: Local Development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT token dari POST /api/auth/google"
  schemas:
...
    Market:
      type: object
      properties:
        id:
          type: string
          description: conditionId dari Polymarket
          example: "0x1234abcd..."
        polymarketId:
          type: string
          example: "market-slug-123"
        slug:
          type: string
          example: "will-btc-reach-100k-2026"
        question:
          type: string
          example: "Will Bitcoin reach $100K by end of 2026?"
        description:
          type: string
          nullable: true
        category:
          type: string
          enum: [Trending, New, Politics, Sports, Crypto]
          example: "Crypto"
        tags:
          type: array
          items:
            type: string
          example: ["bitcoin", "crypto", "price"]
        outcomes:
          type: array
          items:
            type: string
          example: ["Yes", "No"]
        outcomePrices:
          type: object
          description: Probabilitas per outcome (0.0 – 1.0)
          example:
            Yes: 0.65
            No: 0.35
        tokens:
          type: array
          items:
            type: object
            properties:
              token_id:
                type: string
              outcome:
                type: string
        volume:
          type: string
          example: "1250000.50"
        volume24h:
          type: string
          example: "45000.00"
        liquidity:
          type: string
          example: "320000.00"
        active:
          type: boolean
          example: true
        closed:
          type: boolean
          example: false
        imageUrl:
          type: string
          nullable: true
        endDate:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    Transaction:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
          enum: [DEPOSIT, WITHDRAWAL, TRADE, WALLET_CONNECT]
        amount:
          type: string
          nullable: true
          example: "500.000000"
        description:
          type: string
          example: "Deposit via CARD"
        status:
          type: string
          enum: [PENDING, COMPLETED, REJECTED]
        metadata:
          type: object
          nullable: true
        createdAt:
          type: string
          format: date-time

    # ── Portfolio ────────────────────────────────────────────────────────────
    Position:
      type: object
      properties:
        id:
          type: string
        market:
          type: object
          properties:
            id:
              type: string
            question:
              type: string
            category:
              type: string
            imageUrl:
              type: string
              nullable: true
        outcome:
          type: string
          example: "Yes"
        shares:
          type: string
          example: "100.000000"
        avgPrice:
          type: string
          example: "0.600000"
        currentPrice:
          type: string
          example: "0.650000"
        value:
          type: string
          example: "65.00"
        pnl:
          type: string
          example: "5.00"
        color:
          type: string
          nullable: true
        isSimulated:
          type: boolean
    Portfolio:
      type: object
      properties:
        balance:
          type: string
          example: "950.000000"
        totalValue:
          type: string
          example: "65.00"
        pnl:
          type: string
          example: "5.00"
        pnlPercent:
          type: string
          example: "8.33"
        positionCount:
          type: integer
          example: 3
        positions:
          type: array
          items:
            $ref: '#/components/schemas/Position'
        polymarketPositions:
          type: array
          description: Posisi real dari Polymarket (jika user connect wallet)
          items:
            type: object

# ─────────────────────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /api/user/portfolio:
    get:
      tags: [User]
      summary: Get Portfolio
      description: |
        Posisi trading user + statistik portfolio (total value, P&L).
        Jika user memiliki wallet address yang terhubung, juga fetch posisi real dari Polymarket.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Portfolio data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Portfolio'
        '401':
          description: Unauthorized
        '404':
          description: User not found
  /api/user/activity:
    get:
      tags: [User]
      summary: Get Activity / Transaction History
      description: Histori transaksi user — deposit, withdrawal, trade, wallet connect.
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: type
          in: query
          schema:
            type: string
            enum: [DEPOSIT, WITHDRAWAL, TRADE, WALLET_CONNECT]
          description: Filter berdasarkan tipe transaksi
      responses:
        '200':
          description: Daftar transaksi
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  total:
                    type: integer
        '401':
          description: Unauthorized
  /api/user/deposit:
    post:
      tags: [User]
      summary: Deposit (Demo Mode)
      description: |
        Tambah balance demo user.
        **Ini adalah simulasi — tidak ada transaksi uang nyata.**
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount:
                  type: number
                  minimum: 0.01
                  maximum: 100000
                  example: 500
                method:
                  type: string
                  enum: [card, bank]
                  default: card
      responses:
        '200':
          description: Deposit berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Deposit successful (Demo Mode)"
                  newBalance:
                    type: string
                    example: "1500.000000"
                  amount:
                    type: number
                    example: 500
        '400':
          description: Validation error
        '401':
          description: Unauthorized
  /api/user/withdraw:
    post:
      tags: [User]
      summary: Withdraw (Demo Mode)
      description: |
        Kurangi balance demo user.
        **Ini adalah simulasi — tidak ada transfer uang nyata.**
        Akan gagal jika balance tidak mencukupi.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount:
                  type: number
                  minimum: 0.01
                  example: 200
                walletAddress:
                  type: string
                  nullable: true
                  example: "0x742d35Cc6634C0532925a3b8D4C9B5D1b2c3D4E5"
      responses:
        '200':
          description: Withdraw berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Withdrawal successful (Demo Mode)"
                  newBalance:
                    type: string
                    example: "800.000000"
                  amount:
                    type: number
                    example: 200
        '400':
          description: Insufficient balance atau validation error
        '401':
          description: Unauthorized
  /api/user/watchlist:
    get:
      tags: [User]
      summary: Get Watchlist
      description: Daftar market yang di-bookmark user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Daftar market di watchlist
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Market'
        '401':
          description: Unauthorized
  /api/user/watchlist/{marketId}:
    post:
      tags: [User]
      summary: Toggle Watchlist
      description: |
        **Toggle** — tambah ke watchlist jika belum ada, hapus jika sudah ada.
        Response `saved: true` berarti ditambahkan, `saved: false` berarti dihapus.
      security:
        - BearerAuth: []
      parameters:
        - name: marketId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Toggle berhasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  saved:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Added to watchlist"
        '401':
          description: Unauthorized
        '404':
          description: Market not found